clean up
intellij auto cleanup tool

+++ src/main/java/com/ck4911/quest/QuestCalibration.java
@@ -70,7 +70,7 @@ public class QuestCalibration {
 
                   calculatedOffsetToRobot =
                       calculatedOffsetToRobot
-                          .times((double) calculateOffsetCount / (calculateOffsetCount + 1))
+                          .times(calculateOffsetCount / (calculateOffsetCount + 1))
                           .plus(offset.div(calculateOffsetCount + 1));
                   calculateOffsetCount++;
                   Logger.recordOutput("QuestCalibration/CalculatedOffset", calculatedOffsetToRobot);
